implemented epoch timestamp conversion for messages stream
